[gbinder] Implemented service polling. JB#43524
Old servicemanager protocol didn't implement registration
notifications. Those have to be implemented with polling.
Note that polling will be active as long as at least one
listener is registered, i.e. it's important to unregister
the listener when you don't need it.

The polling interval is 2 seconds.

It doesn't affect hwservicemanager implementation which relies
on notifications, no polling there.