[gbinder] Added servicemanager presence API. JB#42956
This function allows to wait for servicemanager synchronously,
blocking the event loop in the process (not recommended):

  gbinder_servicemanager_wait()

These two allow to follow the presence state of servicemanager
without blocking the event loop:

  gbinder_servicemanager_is_present()
  gbinder_servicemanager_add_presence_handler()

Note that services need to re-add their names to servicemanager
after it has restarted. This can be done by the presence handler.