[gbinder] Allow GBinderClient to use more than one interface. JB#49360
Or more likely, multiple versions of the same interface. For example,
first N transaction codes may be associated with version 1.0 of the
interface, the next M codes with version 1.1 and so on.

The following new functions have been introduced:

  gbinder_client_new2
  gbinder_client_interface2
  gbinder_client_new_request2

Note that a similar functionality could be achieved by allocating
a separate client for each version of the interface. This is just
a more convenient way of doing pretty much the same thing.

The old API still works, just assumes the same interface for all
transaction codes.