[gbinder] Wait until reply is handled. JB#52557
The driver sends us back one of three transaction completion messages
BR_TRANSACTION_COMPLETE, BR_DEAD_REPLY or BR_FAILED_REPLY after the
target has finished processing BC_REPLY. It's especially important
when the reply carries an object reference. That gives the other side
a chance to reference that object before we drop our reference to it
(which may be the last one, which would cause the object to be freed
and its handle invalidated).