Adapt the protocol change to the add service command
The binder calls to add a service pretty much remains the same in
a single transaction in Android 11:
```
  ...

  data.writeUtf8AsUtf16(name);
  data.writeStrongBinder(service);
  data.writeBool(allowIsolated);
  data.writeInt32(dumpPriority);
  remote()->transact(::android::IBinder::FIRST_CALL_TRANSACTION + 2 /* addService */, data, &_aidl_reply);
  ...

```

However, Android framework requires an additional field `stability`
when reading a strong binder.
http://aospxref.com/android-11.0.0_r21/xref/frameworks/native/libs/binder/Parcel.cpp#185

The `stability` field is crucial and the operation would fail
if it is missing in a binder transaction when adding a service in
Android.
```
09-07 08:37:58.774    20    20 E servicemanager: Can only set known stability, not 0.
```