[gbinder] Avoid using sync transact method for registerForNotifications. SX#APPSUPPORT-189
Using a sync transaction will deadlock if the service is added from the same process.
We use this to check if a gbinder proxy has been registered on the "other side", in order
to avoid race conditions.
Unfortunately with this change we do not know if registerForNotifications has failed or not
except by checking the logs. I'm open to suggestions to solve this differently.